requirements.txt
astroid==2.11.5
attrs==21.4.0
autopep8==1.6.0
bleach==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
commonmark==0.9.1
coverage==6.3.3
cryptography==37.0.2
dill==0.3.4
docutils==0.18.1
idna==3.3
importlib-metadata==4.11.3
iniconfig==1.1.1
isort==5.10.1
jeepney==0.8.0
jsonmerge==1.8.0
jsonschema==4.5.1
keyring==23.5.0
lazy-object-proxy==1.7.1
mccabe==0.7.0
packaging==21.3
pkginfo==1.8.2
platformdirs==2.5.2
pluggy==1.0.0
py==1.11.0
pycodestyle==2.8.0
pycparser==2.21
Pygments==2.12.0
pylint==2.13.9
pyparsing==3.0.9
pyrsistent==0.18.1
pyserial==3.5
pytest==7.1.2
pytest-cov==3.0.0
pytest-mock==3.7.0
readme-renderer==35.0
requests==2.27.1
requests-toolbelt==0.9.1
rfc3986==2.0.0
rich==12.4.1
SecretStorage==3.3.2
six==1.16.0
toml==0.10.2
tomli==2.0.1
twine==4.0.0
urllib3==1.26.9
webencodings==0.5.1
wrapt==1.14.1
zipp==3.8.0
wheel
setuptools
twine
